requirements.txt
opencv-python-headless
ipython
streamlink
twitchAPI
torch
torchvision
python-telegram-bot
pandas
seaborn
matplotlib>=3.2.2
Pillow>=7.1.2
psutil  
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
thop>=0.1.1 
tqdm>=4.64.0
